+++ src/components/Dropdown/index.ts
@@ -3,6 +3,4 @@ import { Person } from '../../types/Person';
 export interface Props {
   users: Person[];
   setSelectedUser: (value: Person | null) => void;
-  query: string;
-  saveQuery({ newQuery }: { newQuery: string }): void;
 }
